v1.5.2.2

+++ beelinegps.ggi.pas
@@ -3,12 +3,16 @@
 function InstallWork: string;
 begin
   // Do install tasks here.
-  if FileExists(GEOGET_DATADIR+'\gpxBeeline.gge.pas') then
-    DeleteFile(GEOGET_DATADIR+'\gpxBeeline.gge.pas');
-  if FileExists(GEOGET_SCRIPTDIR+'\gpxBeeline.gge.pas') then
-    DeleteFile(GEOGET_SCRIPTDIR+'\gpxBeeline.gge.pas');
-  if FileExists(GEOGET_SCRIPTDIR+'\export\gpxBeeline.gge.pas') then
-    DeleteFile(GEOGET_SCRIPTDIR+'\export\gpxBeeline.gge.pas');
+
+  {uklid po starsich verzich}
+  if FileExists(GEOGET_DATADIR+'\gpxBeeline.gge.pas') then DeleteFile(GEOGET_DATADIR+'\gpxBeeline.gge.pas');
+  if FileExists(GEOGET_SCRIPTDIR+'\gpxBeeline.gge.pas') then DeleteFile(GEOGET_SCRIPTDIR+'\gpxBeeline.gge.pas');
+  if FileExists(GEOGET_SCRIPTDIR+'\export\gpxBeeline.gge.pas') then DeleteFile(GEOGET_SCRIPTDIR+'\export\gpxBeeline.gge.pas');
+
+  {changelog}
+  if FileExists(GEOGET_SCRIPTDIR + '\BeelineGPS\BeelineGPS.changelog.txt') then ShowLongMessage('Changelog', FileToString(GEOGET_SCRIPTDIR + '\BeelineGPS\BeelineGPS.changelog.txt'));
+
+  Result := '';  // probehlo bez chyby
 end;
 
 function UninstallWork: string;
